Finished writing ReadGrid function so that suntans works in the following way: 1) Run and create the parallel grid with mpirun -np 3 sun -g 2) Solve for the physics with mpirun -np 3 sun -s 3) Without the -g flag, the function reads the previous files using the ReadGrid function.
Also changed a few lines in Geometry so that xi != nan.   This
was causing errors upon attempting to read in xi when it was nan.

Changed verbose output of "Reading..." and "Writing..." in
OutputData() and ReadGrid() so that they print out the file
being written to/read from on each processor.