Added ComputeTangentialVelocity(gridT *grid, physT *phys), which is needed for advection of horizontal velocity (used to be computed for outputting data only).
Added ComputeTraceBack, which computes the xd,yd, and zd points of
the Lagrangian traceback.  Right now it is a simple xd=x-u*dt.

Added FindNearestEdge, which finds the nearest edge as a member of
the nearestedges array.  This reduces the search time required by
a factor of Ne/Nnearestedges, since the entire grid does not have to
be searched.  This was not the origingal intention of the nearestedges
arrays but turned out to work well!

Added Quadratic, which performs a quadratic interpolation in the
vertical.  As of now, vertical interpolation is performed with
this function to the level of the traceback, and then Kriging
interpolation with linear drift is performed with the Kriging
function.

Added Kriging.  This function computes the value of the horizontal
normal to a face given the normal components at the other edges
in the nearestedges array.

Added InterpolateEdge, which computes the traceback, determines i
f it is within the domain, and then interpolates with Kriging.

FindNearestPlane was added but was not used since the interpolation is
not done using the cells around the traceback point but from the
departure point.

To do: precompute the kriging matrix and its inverse, and more
accurately compute the vertical interpolation.  Quadratic is very
inaccurate because it does not take into account different vertical
grid spacings at the fs and bottom.