Fixed a very strange bug in report.c.  For some reason the code segment
while(argv[i][j]!='=')
  str[j-js]=argv[i][j++];

Was incrementing j in j-js before using it when compiled
with the intel compiler.  For example, with the gcc compiler,
when j starts at 2 and js starts at 2, then the first
entry should execute

str[0]=argv[i][2];
and then j will update after it is used to 3.

But with icc upon entry into the loop it interprets it
as
str[1]=argv[i][2]
because it is updating j first...Is this an icc bug?

The easiest fix was to add j++ after the update, i.e.

while(argv[i][j]!='=') {
  str[j-js]=argv[i][j];
  j++;
}

This works with gcc and icc, but of course is not as clean.