Fixed out-of-bounds errors in phys.c:
1) Changed phys->uold to phys->wtmp in GuessQ since Nk+1 vertical levels
are allocated in w variables but only Nk in u variables

2) Changed the k limit from grid->Nk[i] to grid->Nke[ne] in
ComputeQSource since the fluxes below nke are zero anyway and do not
contribute to the source term

3) In OperatorQ added the if statement:
if(grid->ctop[i]<grid->Nk[i]-1) {
which ensures that no out-of-bounds references are made when there is
only 1 vertical level.

4) In UpdateScalars changed the line
for(k=0;k<grid->Nkc[ne];k++)
to
for(k=0;k<grid->Nke[ne];k++)
since it produces an out-of-bounds error.  Going to Nkc looks for
scalar values beneath the bottom in neighboring cells and is not
necessary.  These fluxes are zero anyway.