Made several changes, including:
1) Added a fast preconditioner which approximates the Poisson equation
as p_zz = S^* since p_xx is approximately zero.  This requires that
the preconditioner perform tridiagonal inversions in order to obtain
the approximate solution.  The approximate solution can actually be
obtained by using the preconditioner as
Preconditioner(source,q,coefs,grid,phys,prop);
since this function performs the operation
q = M^(-1) src
This preconditioner can be found in the paper by Marshall, et al. (JGR 102,C3,pp5753-5766)

To use this preconditioner, set qprecond=2 in suntans.dat.  If qprecond=1, then
the preconditioner is a diagonal preconditioner, while if qprecond=0, then
no preconditioner is used.  qprecond=2 is roughly 5 times faster than qprecond=1.

2) Using the turbmodel variable, space is allocated for qT,lT,Cn_q,Cn_l only if
turbmodel is set to 1. Note that these variables are only written to the
restart file if turbmodel=1, so if a run is restarted it must have the same
turbmodel value set otherwise it will read past EOF and crash.

Note that the if statement "if(prop->turbmodel)" is used in:
AllocatePhysicalVariables
FreePhysicalVariables
InitializePhysicalVariables
ReadPhysicalVariables (from restart file)
OutputData
EddyViscosity

3) Fixed a bug in ComputeQSource which was originally fetching out-of-bounds
memory for variable bathymetry.  Original code fragment:
      for(k=grid->ctop[i];k<grid->Nk[i];k++) {
	ap[k] = 0.5*(phys->u[ne][k]+fabs(phys->u[ne][k]));
	am[k] = 0.5*(phys->u[ne][k]-fabs(phys->u[ne][k]));
      }

New code fragment:
      for(k=grid->ctop[i];k<grid->Nke[ne];k++) {
	ap[k] = 0.5*(phys->u[ne][k]+fabs(phys->u[ne][k]));
	am[k] = 0.5*(phys->u[ne][k]-fabs(phys->u[ne][k]));
      }
      for(k=grid->Nke[ne];k<grid->Nkc[ne];k++) {
	ap[k] = 0;
	am[k] = 0;
      }

The old code fragment fetched out-of-bounds data since u[ne][grid->Nk[i]-1]
will be out-of-bounds if grid->Nk[ne]<grid->Nk[i] which it usually is
on at least one column edge for variable bathymetry.

4) Fixed the reporting of residual convergence so that a warning is reported if
residual>epsilon after both the pressure and free-surface iterations.  This will
only be printed out if VERBOSE>2.

5) In the ConditionQ function changed the i loop to be from celldist[0] to celldist[1]
rather than to celldist[2] since this is not necessary.  Also removed the kmax
variable in the k loop since the k loop only needs to go to Nke[ne] on each face.

Added the if statement  if(grid->ctop[i]<grid->Nk[i]-1) in several areas of
the pressure solution to prevent out-of-bounds errors from appearing when only
1 vertical layer is present.