Added the following lines of code to OperatorQ and Preconditioner:
To OperatorQ, added the line which applies the operator when
the number of cells in the vertical is 1.  This was preventing the
pressure solver from converging because the values of q for single-layer
cells were not changing with the iteration and hence the residual
would asymptote to a value well above the residual.
     } else
       y[i][grid->ctop[i]]-=2.0*coef[i][grid->ctop[i]]*x[i][grid->ctop[i]];

Also added the same line to the preconditioner, which divides by
the diagonal coefficient when there is only one layer in the vertical.

     } else
       xc[i][grid->ctop[i]]=-0.5*x[i][grid->ctop[i]]/coef[i][grid->ctop[i]];