Changed thetaQ to 1.0 instead of 0.75 since 0.75 is not always stable.
Added the ability to specify incoming turbulence quantities.  Right now
they are set so that there is no gradient on the turbulent quantitites
for boundary_flag values of open or not open.  In order to change this
the code must be changed in the loops that loop over edgedist[4] to
edgedist[5].

Also changed the (completely ad-hoc) initialization of the turbulent
length scale so that it grows for the first 30 time steps instead of
the first 3.