Made changes to account for obtuse triangles.  The original CorrectVoronoi function does not account for bad triangles because dg[] is not necessarily small for obtuse triangles. For example, a triangle with an angle of 120 degrees is grossly degenerate, but the corresponding dg[] on the opposite Voronoi side may be quite large.  Therefore, changes have been made so that now it is better to use CorrectVoronoi=-1 in suntans.dat with the corresponding value of VoronoiRatio set as the maximum allowable angle in degrees. The original function is used when CorrectVoronoi=1 for backwards compatibility.
1) Added CorrectAngles function that is used when CorrectVoronoi==-1.
In this case, VoronoiRatio is taken as the maximum angle allowed in
any of the triangles.  Typical values should be around 85 so that
triangles with angles>85 degrees are corrected, i.e. the Voronoi
point is moved to the centroid.

This affects the calculation of def[]  and xe[], ye[].  def[]
is defined as the length of the line that passes through the
(possibly corrected) Voronoi point and is perpendicular to a
triangle edge.  For non-corrected triangles, def[] corresponds
to the length of the perpendicular bisector and passes through
xe,ye, which is the edge midpoint.  Upon correcting, the point
xe,ye no longer corresponds to the midpoint but instead corresponds
to the intersection of the Delaunay edge with the Voronoi edge.

Note that def[] is used primarly for calculation of uc, and vc,
as well as advection of momentum, while xe,ye is only used for
interpolation of values to the faces.  This is reflected by
changes in phys.c.

2) Added VoronoiStats function which prints out grid statistics
before and after correction. Running with -vvv prints out a histogram
with NUMDGHIST bins as defined in grid.h.