- Added send/recv of initial velocity field in InitializePhysicalVariables to avoid discontinuities in initial conditions at interproc boundaries.
- Fixed momentum advection to only use boundary velocity data if the flow is
into the domain for type 2 boundaries.

- Changed type 3 momentum advection to always perform first-order upwinding
from boundary data to the interior.  This ensures that outflow flux is
obtained from the interior and inflow flux is obtained from bc's.

- Fixed vertical advection of momentum so that the top cell is computed in
a non-conservative way.  The previous method was not conservative.  The change
only required addition of k=grid->ctop[i] to the vertical advection loops.

- Used phys->utmp2 for advection of vertical momentum since utmp2 is the old
velocity at time step n, while phys->u is u*.  This ensures that the velocity
field used to compute advection of w is divergence-free (since time step n
is divergence-free, while time step n for w and time step * for u is not.